Split each step into separate file to minimize conflicts
