+++ src/App.css
@@ -102,157 +102,3 @@ a:link {
   text-align: center;
   padding: 10px;
 }
-
-/* NextStepButton */
-
-.NextStepButton {
-  color: white;
-  display: flex;
-  margin-bottom: 25px;
-  width: 75%;
-  height: 10vh;
-  background-color: #f88920;
-  text-transform: uppercase;
-  letter-spacing: 1px;
-  justify-content: center;
-  align-items: center;
-  font-family: 'Lato';
-  font-weight: 300;
-  font-size: 20px;
-}
-
-.NextStepButton--disabled {
-  background-color: #EDEDED;
-}
-
-.NextStepButton > a {
-  text-decoration: none;
-}
-
-.NextStepButton > a:visited {
-  color: white;
-}
-
-/* DonateFormPage1 */
-
-.DonateFormPage1__wrapper {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-self: center;
-  justify-content: space-around;
-  align-items: center;
-  height: 60vh;
-}
-
-.DonateFormPage1__donationWrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  align-items: center;
-  align-content: center;
-  height: 50vh;
-}
-
-.DonateFormPage1__inputWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-}
-
-.DonateFormPage1__input {
-  display:flex;
-  height: 7vh;
-  width: 60%;
-  border: 1px solid #DADCDE;
-  box-shadow: 0 2px 4px 0 rgba(0,0,0,0.05);
-  border-radius: 2px;
-  padding: 0px 0px 0px 16px;
-  font-size: 16px;
-  line-height: 19px;
-  font-family: 'Lato', arial, sans-serif;
-  box-sizing: border-box;
-}
-
-.DonateFormPage1__input:focus {
-  outline: none;
-}
-
-.DonateFormPage1__CTA {
-  color: white;
-  display: flex;
-  margin: 10px;
-  width: 75%;
-  height: 10vh;
-  background-color: #f88920;
-  font-weight: 300;
-  letter-spacing: 1px;
-  justify-content: center;
-  align-items: center;
-  text-transform: uppercase;
-  font-size: 20px;
-}
-
-.DonateFormPage1__CTA > a {
-  text-decoration: none;
-}
-
-.DonateFormPage1__CTA > a:visited {
-  color: white;
-}
-
-/* Donate form page 2 */
-
-.DonateFormPage2__wrapper {
-  display: flex;
-  width: 90vw;
-  max-width: 350px;
-  flex-direction: column;
-  align-self: center;
-  justify-content: space-around;
-  align-items: center;
-  height: 90vh;
-}
-
-.DonateFormPage2__content {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 90vw;
-  font-family: 'Lato';
-  font-weight: 400;
-  font-size: 18px;
-  color: #424242;
-  height: 40vh;
-}
-
-.DonateFormPage2__content > img {
-  margin: auto;
-  width: 50%;
-  height: 50%;
-}
-
-
-.DonateFormPage2__CTA {
-  color: white;
-  display: flex;
-  margin: 10px;
-  width: 75%;
-  height: 10vh;
-  background-color: #f88920;
-  font-weight: 300;
-  letter-spacing: 1px;
-  justify-content: center;
-  align-items: center;
-  align-self: flex-end;
-  text-transform: uppercase;
-  font-size: 20px;
-}
-
-.DonateFormPage2__CTA > a {
-  text-decoration: none;
-}
-
-.DonateFormPage2__CTA > a:visited {
-  color: white;
-}
