+++ src/constants.js
@@ -1,0 +1,2 @@
+export const NUMBER_OF_YEARS = 10;
+export const DEFAULT_DONATION_REPEAT = 'monthly';
