+++ src/utils.js
@@ -1,0 +1,3 @@
+export const humanize = x => {
+  return x.toFixed(2).replace(/\.?0*$/, '');
+};
